Updated utility classes and organization - Moved back class `Utils.Compatibility.ImprovedInputHandler` to `Utils.ImprovedInputHandler` - Updated class `Utils.InputHandler`:   - Now always tries to register an IIC-compatible keybind upon registering a new `Keybind`     - Ignores any `FileNotFoundException`-related errors for logging purposes     - Uses the same technique for determining if IIC is, indeed, enabled at startup - Updated class `Utils.OptionBuilder`:   - Changed version text to show up closer to the title, and more to the right   - Fixed combo box (and its label) all over the place when placed on the menu   - Method `AddPadding(Vector2)` now subtracts from the internal vector for more intuitive usage
